+++ tests/no-markers/main.cpp
@@ -15,8 +15,16 @@ void foo(int param) {
   // LCOV_EXCL_STOP
 }
 
+void bar(int param) {  // GCOVR_EXCL_FUNCTION
+  if (param) {
+     std::cout << "param is true" << std::endl;
+  } else {
+     std::cout << "param is false" << std::endl;
+  }
+}
 
 int main(int argc, char* argv[]) {
   foo(argc-1);
+  bar(argc-1);
   return 0;
 }